Fix ocr package crashing
